+++ src/VirtoCommerce.PageBuilderModule.Web/page-builder/package.json
@@ -24,9 +24,9 @@
     "@types/node": "^20.10.5",
     "@typescript-eslint/eslint-plugin": "^7.4.0",
     "@typescript-eslint/parser": "^7.4.0",
-    "@vc-shell/api-client-generator": "^1.1.52",
-    "@vc-shell/release-config": "^1.1.52",
-    "@vc-shell/ts-config": "^1.1.52",
+    "@vc-shell/api-client-generator": "^1.2.2",
+    "@vc-shell/release-config": "^1.2.2",
+    "@vc-shell/ts-config": "^1.2.2",
     "@vitejs/plugin-vue": "^5.2.3",
     "@volar/typescript": "^2.4.14",
     "@vue/eslint-config-prettier": "^9.0.0",
@@ -55,8 +55,8 @@
     "vue-tsc": "^2.2.10"
   },
   "dependencies": {
-    "@vc-shell/config-generator": "^1.1.52",
-    "@vc-shell/framework": "^1.1.52",
+    "@vc-shell/config-generator": "^1.2.2",
+    "@vc-shell/framework": "^1.2.2",
     "@vueuse/core": "^10.7.1",
     "@vueuse/integrations": "^10.7.1",
     "cross-spawn": "^7.0.3",
