+++ src/VirtoCommerce.PageBuilderModule.Web/page-builder/src/modules/page-builder/components/index.ts
@@ -1,3 +1,6 @@
 import PageStatus from "./pageStatus.vue";
+import CardHeader from "./cardHeader.vue";
+import SwitchRow from "./switchRow.vue";
+import Tag from "./tag.vue";
 
-export { PageStatus };
+export { PageStatus, CardHeader, SwitchRow, Tag };
